+++ packages/containers-shared/tests/utils.test.ts
@@ -1,19 +1,12 @@
-import {
-	execFile,
-	execFileSync,
-	spawn,
-	type ChildProcess,
-} from "node:child_process";
+import { execFile, spawn, type ChildProcess } from "node:child_process";
 import { EventEmitter } from "node:events";
 import { existsSync } from "node:fs";
 import { release } from "node:os";
 import { afterEach, beforeEach, describe, it, vi } from "vitest";
 import {
-	checkExposedPorts,
 	containerPrivilegesAllowed,
 	verifyDockerInstalled,
 } from "./../src/utils";
-import type { ContainerDevOptions } from "../src/types";
 
 type DockerExecFile = (
 	file: string,
@@ -23,7 +16,6 @@ type DockerExecFile = (
 ) => ChildProcess;
 
 const dockerExecFile = execFile as DockerExecFile;
-let docketImageInspectResult = "0";
 
 vi.mock("node:child_process");
 vi.mock("node:fs");
@@ -31,59 +23,6 @@ vi.mock("node:os", async (importOriginal) => ({
 	...(await importOriginal<typeof import("node:os")>()),
 	release: vi.fn(),
 }));
-
-vi.mock("../src/inspect", async (importOriginal) => {
-	const mod: object = await importOriginal();
-	return {
-		...mod,
-		dockerImageInspect: () => docketImageInspectResult,
-	};
-});
-
-const containerConfig = {
-	dockerfile: "",
-	class_name: "MyContainer",
-} as ContainerDevOptions;
-describe("checkExposedPorts", () => {
-	beforeEach(() => {
-		docketImageInspectResult = "1";
-		vi.mocked(execFileSync).mockReset();
-	});
-
-	it("should not error when some ports are exported", async ({ expect }) => {
-		docketImageInspectResult = "1";
-		await expect(
-			checkExposedPorts("docker", containerConfig)
-		).resolves.toBeUndefined();
-	});
-
-	it("should error, with an appropriate message when no ports are exported", async ({
-		expect,
-	}) => {
-		docketImageInspectResult = "0";
-		await expect(checkExposedPorts("docker", containerConfig)).rejects
-			.toThrowErrorMatchingInlineSnapshot(`
-				[Error: The container "MyContainer" does not expose any ports. In your Dockerfile, please expose any ports you intend to connect to.
-				For additional information please see: https://developers.cloudflare.com/containers/local-dev/#exposing-ports.
-				]
-			`);
-	});
-
-	it("identifies the named image that does not expose ports", async ({
-		expect,
-	}) => {
-		docketImageInspectResult = "0";
-		await expect(
-			checkExposedPorts("docker", {
-				...containerConfig,
-				image_name: "app",
-			})
-		).rejects.toThrow(
-			'The container "MyContainer.app" does not expose any ports'
-		);
-	});
-});
-
 describe("containerPrivilegesAllowed", () => {
 	let commandError: Error | null;
 	let rawResponse: string | undefined;